fix(ceiling): move operator permission ceiling to src/shared so agent code can import it
operator_tools.py (shipped in the agent container, which has no src/host) imported the ceiling from src.host.permissions -> ModuleNotFoundError at runtime. Move _OPERATOR_PERMISSION_CEILING + clamp_to_operator_ceiling to src/shared/operator_ceiling.py; host/permissions re-exports them so server.py is unchanged. (Final-review CI regression; respects the src.agent->src.host trust-boundary guardrail.)
